# apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "SolidCouch Geoindex",
    "description": "This is a geoindex for Solid apps. Read more at https://github.com/solidcouch/geoindex"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/.well-known/openid-configuration": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/jwks": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/profile/card": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/inbox": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/ld+json": {
              "schema": {
                "$ref": "#/components/schemas/notification"
              }
            }
          }
        }
      }
    },
    "/query": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "object",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "notification": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "string",
            "const": "https://www.w3.org/ns/activitystreams"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "Create",
              "Update",
              "Delete"
            ]
          },
          "actor": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "Person"
              },
              "id": {
                "type": "string",
                "format": "uri"
              }
            },
            "required": [
              "type",
              "id"
            ]
          },
          "object": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Document",
                  "Place"
                ]
              },
              "id": {
                "type": "string",
                "format": "uri"
              }
            },
            "required": [
              "type",
              "id"
            ]
          }
        },
        "required": [
          "@context",
          "type",
          "actor",
          "object"
        ],
        "additionalProperties": false
      }
    }
  }
}